+++ examples/reload-feature/src/main.rs
@@ -6,7 +6,7 @@ use lib::*;
 #[cfg(feature = "reload")]
 #[hot_lib_reloader::hot_module(dylib = "lib")]
 mod hot_lib {
-    hot_functions_from_file!("../lib/src/lib.rs");
+    hot_functions_from_file!("lib/src/lib.rs");
 }
 
 fn main() {