+++ tests/lib-loader-test.rs
@@ -1,8 +1,8 @@
 mod common;
 
-#[hot_lib_reloader::hot_module(dylib = "lib_for_testing")]
+#[hot_lib_reloader::hot_module(dylib = "lib_for_testing", file_watch_debounce = 50)]
 mod hot_lib {
-    hot_functions_from_file!("./lib_for_testing/src/lib.rs");
+    hot_functions_from_file!("tests/lib_for_testing/src/lib.rs");
 
     #[lib_change_subscription]
     pub fn subscribe() -> hot_lib_reloader::LibReloadObserver {}